Files for 0.6
Multiple things added. Getting ready for a full release here. See my description in this release.

+++ Bash For Windows/rm.py
@@ -1,7 +1,7 @@
 '''
 This file is under the MIT License.
 
-Copyright 2019 Jeremiah Haven
+Copyright 2019-2020 Jeremiah Haven
 
 Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files 
 (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, 
@@ -19,14 +19,62 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 # Libraries
 import os
+import shutil
 import ls
+import systemvariables
 
 # Main Function. First get a listing of the directory from the ls script
-def remove(file):
+def remove(args):
     directory = ls.list()
-    
-    # If it exists, remove it. If not, display a message
-    if(any(file in s for s in directory)):
-        os.remove(file)
+    i = 0
+    m = 0
+    n = 0
+    if("-r" in args):
+        lastIndex = len(args) - 1
+        if(args[0] == "-r"):
+            args.pop(0)
+            for j in args:
+                file = args[i]
+                if(any(file in s for s in directory)):
+                    # Make sure it's' a directory
+                    if(os.path.isdir(file) == True):
+                        if(os.getcwd() == systemvariables.HOME):
+                            if(file == "Documents"):
+                                force = input("rm: remove write-protected directory '" + file + "'? ")
+                                if((force == "y") or (force == "Y")):
+                                    shutil.rmtree(file)
+                            else:
+                                shutil.rmtree(file)
+                        else:
+                            shutil.rmtree(file)
+                    else:
+                        os.remove(file)
+                else:
+                    print("rm: cannot remove '" + file + ": No such file or directory")
+                i += 1
+        elif(args[lastIndex] == "-r"):
+            args.pop(lastIndex)
+            for j in args:
+                file = args[i]
+                if(any(file in s for s in directory)):
+                    # Make sure it isn't a directory
+                    if(os.path.isdir(file) == True):
+                        shutil.rmtree(file)
+                    else:
+                        os.remove(file)
+                else:
+                    print("rm: cannot remove '" + file + ": No such file or directory")
+                i += 1
     else:
-        print("rm: The file", file, "does not exist so not deleting!")
+        for j in args:
+            file = args[i]
+            # If it exists, remove it. If not, display a message
+            if(any(file in s for s in directory)):
+                # Make sure it isn't a directory
+                if(os.path.isfile(file) == True):
+                    os.remove(file)
+                else:
+                    print("rm: cannot remove '" + file + ": Is a directory")
+            else:
+                print("rm: cannot remove '" + file + ": No such file or directory")
+            i += 1
